etcd-runner: move root cmd into command package
this allows easier sharing of global variable for sub commands.
