+++ tools/functional-tester/etcd-runner/command/help.go
@@ -14,7 +14,7 @@
 
 // copied from https://github.com/rkt/rkt/blob/master/rkt/help.go
 
-package main
+package command
 
 import (
 	"bytes"
